+++ app/scenes/Settings/Security.tsx
@@ -1,27 +1,21 @@
 import { debounce } from "lodash";
 import { observer } from "mobx-react";
-import { CloseIcon, PadlockIcon } from "outline-icons";
+import { PadlockIcon } from "outline-icons";
 import { useState } from "react";
 import * as React from "react";
 import { useTranslation, Trans } from "react-i18next";
-import styled from "styled-components";
-import Button from "~/components/Button";
 import ConfirmationDialog from "~/components/ConfirmationDialog";
-import Fade from "~/components/Fade";
-import Flex from "~/components/Flex";
 import Heading from "~/components/Heading";
-import Input from "~/components/Input";
 import InputSelect from "~/components/InputSelect";
-import NudeButton from "~/components/NudeButton";
 import Scene from "~/components/Scene";
 import Switch from "~/components/Switch";
 import Text from "~/components/Text";
-import Tooltip from "~/components/Tooltip";
 import env from "~/env";
 import useCurrentTeam from "~/hooks/useCurrentTeam";
 import useStores from "~/hooks/useStores";
 import useToasts from "~/hooks/useToasts";
 import isCloudHosted from "~/utils/isCloudHosted";
+import DomainManagement from "./components/DomainManagement";
 import SettingRow from "./components/SettingRow";
 
 function Security() {
@@ -37,15 +31,6 @@ function Security() {
     memberCollectionCreate: team.memberCollectionCreate,
     inviteRequired: team.inviteRequired,
   });
-
-  const [allowedDomains, setAllowedDomains] = useState([
-    ...(team.allowedDomains ?? []),
-  ]);
-  const [lastKnownDomainCount, updateLastKnownDomainCount] = useState(
-    allowedDomains.length
-  );
-
-  const [existingDomainsTouched, setExistingDomainsTouched] = useState(false);
 
   const authenticationMethods = team.signinMethods;
 
@@ -80,21 +65,6 @@ function Security() {
     },
     [data, saveData]
   );
-
-  const handleSaveDomains = React.useCallback(async () => {
-    try {
-      await auth.updateTeam({
-        allowedDomains,
-      });
-      showSuccessMessage();
-      setExistingDomainsTouched(false);
-      updateLastKnownDomainCount(allowedDomains.length);
-    } catch (err) {
-      showToast(err.message, {
-        type: "error",
-      });
-    }
-  }, [auth, allowedDomains, showSuccessMessage, showToast]);
 
   const handleDefaultRoleChange = React.useCallback(
     async (newDefaultRole: string) => {
@@ -141,42 +111,6 @@ function Security() {
     [data, saveData, t, dialogs, authenticationMethods]
   );
 
-  const handleRemoveDomain = async (index: number) => {
-    const newDomains = allowedDomains.filter((_, i) => index !== i);
-
-    setAllowedDomains(newDomains);
-
-    const touchedExistingDomain = index < lastKnownDomainCount;
-    if (touchedExistingDomain) {
-      setExistingDomainsTouched(true);
-    }
-  };
-
-  const handleAddDomain = () => {
-    const newDomains = [...allowedDomains, ""];
-
-    setAllowedDomains(newDomains);
-  };
-
-  const createOnDomainChangedHandler = (index: number) => (
-    ev: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    const newDomains = allowedDomains.slice();
-
-    newDomains[index] = ev.currentTarget.value;
-    setAllowedDomains(newDomains);
-
-    const touchedExistingDomain = index < lastKnownDomainCount;
-    if (touchedExistingDomain) {
-      setExistingDomainsTouched(true);
-    }
-  };
-
-  const showSaveChanges =
-    existingDomainsTouched ||
-    allowedDomains.filter((value: string) => value !== "").length > // New domains were added
-      lastKnownDomainCount;
-
   return (
     <Scene title={t("Security")} icon={<PadlockIcon color="currentColor" />}>
       <Heading>{t("Security")}</Heading>
@@ -221,64 +155,7 @@ function Security() {
       )}
 
       {!data.inviteRequired && (
-        <SettingRow
-          label={t("Allowed domains")}
-          name="allowedDomains"
-          description={t(
-            "The domains which should be allowed to create new accounts using SSO. Changing this setting does not affect existing user accounts."
-          )}
-        >
-          {allowedDomains.map((domain, index) => (
-            <Flex key={index} gap={4}>
-              <Input
-                key={index}
-                id={`allowedDomains${index}`}
-                value={domain}
-                autoFocus={!domain}
-                placeholder="example.com"
-                required
-                flex
-                onChange={createOnDomainChangedHandler(index)}
-              />
-              <Remove>
-                <Tooltip tooltip={t("Remove domain")} placement="top">
-                  <NudeButton onClick={() => handleRemoveDomain(index)}>
-                    <CloseIcon />
-                  </NudeButton>
-                </Tooltip>
-              </Remove>
-            </Flex>
-          ))}
-
-          <Flex justify="space-between" gap={4} style={{ flexWrap: "wrap" }}>
-            {!allowedDomains.length ||
-            allowedDomains[allowedDomains.length - 1] !== "" ? (
-              <Fade>
-                <Button type="button" onClick={handleAddDomain} neutral>
-                  {allowedDomains.length ? (
-                    <Trans>Add another</Trans>
-                  ) : (
-                    <Trans>Add a domain</Trans>
-                  )}
-                </Button>
-              </Fade>
-            ) : (
-              <span />
-            )}
-
-            {showSaveChanges && (
-              <Fade>
-                <Button
-                  type="button"
-                  onClick={handleSaveDomains}
-                  disabled={auth.isSaving}
-                >
-                  <Trans>Save changes</Trans>
-                </Button>
-              </Fade>
-            )}
-          </Flex>
-        </SettingRow>
+        <DomainManagement onSuccess={showSuccessMessage} />
       )}
       {!data.inviteRequired && (
         <SettingRow
@@ -347,9 +224,5 @@ function Security() {
     </Scene>
   );
 }
-
-const Remove = styled("div")`
-  margin-top: 6px;
-`;
 
 export default observer(Security);
