+++ shared/i18n/locales/en_US/translation.json
@@ -578,6 +578,11 @@
   "We were unable to find the page you’re looking for.": "We were unable to find the page you’re looking for.",
   "No documents found for your search filters.": "No documents found for your search filters.",
   "Search Results": "Search Results",
+  "Allowed domains": "Allowed domains",
+  "The domains which should be allowed to create new accounts using SSO. Changing this setting does not affect existing user accounts.": "The domains which should be allowed to create new accounts using SSO. Changing this setting does not affect existing user accounts.",
+  "Remove domain": "Remove domain",
+  "Add a domain": "Add a domain",
+  "Save changes": "Save changes",
   "Please choose a single file to import": "Please choose a single file to import",
   "Your import is being processed, you can safely leave this page": "Your import is being processed, you can safely leave this page",
   "File not supported – please upload a valid ZIP file": "File not supported – please upload a valid ZIP file",
@@ -715,11 +720,6 @@
   "The server must have SMTP configured to enable this setting": "The server must have SMTP configured to enable this setting",
   "Require invites": "Require invites",
   "Require members to be invited to the workspace before they can create an account using SSO.": "Require members to be invited to the workspace before they can create an account using SSO.",
-  "Allowed domains": "Allowed domains",
-  "The domains which should be allowed to create new accounts using SSO. Changing this setting does not affect existing user accounts.": "The domains which should be allowed to create new accounts using SSO. Changing this setting does not affect existing user accounts.",
-  "Remove domain": "Remove domain",
-  "Add a domain": "Add a domain",
-  "Save changes": "Save changes",
   "Default role": "Default role",
   "The default user role for new accounts. Changing this setting does not affect existing user accounts.": "The default user role for new accounts. Changing this setting does not affect existing user accounts.",
   "Behavior": "Behavior",
